Removed dependency injection of name/version
Now that the main plugin class is a singleton, there’s no need to pass
the name and version to the Admin/Public/Common classes as
dependencies. By removing these dependencies, we save a huge amount of
duplicated code in each class and possibly allow the other classes to
become singletons as well.